requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
faiss-cpu==1.7.4
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
numpy==1.24.3
sentence-transformers==2.2.2
python-dotenv==1.0.0